pyproject.toml
[project]
name = "roadtrip-skills"
version = "0.1.0"
description = "Autonomous multi-agent skill framework for RoadTrip project"
requires-python = ">=3.10"
dependencies = [
	"pyyaml>=6.0",
	"mcp>=1.0.0",
	"youtube-transcript-api>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
